move to inner class
